requirements.txt
Package                       Version
----------------------------- ---------
asgiref                       3.5.2
atomicwrites                  1.4.1
attrs                         21.4.0
certifi                       2022.6.15
cffi                          1.15.1
charset-normalizer            2.1.0
colorama                      0.4.5
coreapi                       2.3.3
coreschema                    0.0.4
cryptography                  37.0.4
defusedxml                    0.7.1
Django                        4.0.6
django-js-asset               2.0.0
django-mptt                   0.13.4
django-templated-mail         1.1.1
djangorestframework           3.13.1
djangorestframework-simplejwt 4.8.0
djoser                        2.1.0
factory-boy                   3.2.1
Faker                         13.15.0
idna                          3.3
inflection                    0.5.1
iniconfig                     1.1.1
itypes                        1.2.0
Jinja2                        3.1.2
MarkupSafe                    2.1.1
oauthlib                      3.2.0
packaging                     21.3
pip                           22.2
pluggy                        1.0.0
py                            1.11.0
pycparser                     2.21
PyJWT                         2.4.0
pyparsing                     3.0.9
pytest                        6.2.5
pytest-base-url               1.4.2
pytest-django                 4.5.2
pytest-factoryboy             2.5.0
pytest-html                   3.1.1
pytest-metadata               2.0.2
pytest-selenium               3.0.0
pytest-variables              1.9.0
python-dateutil               2.8.2
python-decouple               3.6
python3-openid                3.2.0
pytz                          2022.1
requests                      2.28.1
requests-oauthlib             1.3.1
selenium                      3.141.0
setuptools                    58.1.0
six                           1.16.0
social-auth-app-django        4.0.0
social-auth-core              4.3.0
sqlparse                      0.4.2
tenacity                      6.3.1
toml                          0.10.2
tomli                         2.0.1
typing_extensions             4.3.0
tzdata                        2022.1
uritemplate                   4.1.1
urllib3                       1.26.10
